+++ packages/vite-plugin-lxr/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.0] - 2023-05-04
+
+### Changed
+- dev server command: "vite" (requires "lxr.json" file in project root folder)
+- upload command: "vite build --mode upload" (requires "lxr.json" file in project root folder)
+- build only command: "vite build"
+
 ## [0.0.8] - 2022-11-07
 
 ### Fixed